+++ tests/test_coordspace.py
@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 
 import nibabel as nb
@@ -17,7 +18,8 @@ datatype = 'ieeg'
 space = 'fs'
 
 # paths to test files
-bids_root = Path('data')
+cwd = os.getcwd()
+bids_root = Path(cwd) / 'data'
 deriv_root = bids_root / 'derivatives'
 mri_dir = deriv_root / 'freesurfer' / f'sub-{subject}' / 'mri'
 subjects_dir = deriv_root / 'freesurfer'
@@ -36,9 +38,20 @@ _bids_path = BIDSPath(subject=subject, session=session,
 
 
 @pytest.mark.usefixtures('_temp_bids_root')
-@pytest.mark.parametrize('to_coord', ['mri', 'mm', 'tkras', 'm'])
-def test_convert_coordunits(_temp_bids_root, to_coord):
+@pytest.mark.parametrize('to_frame, to_unit', [
+    ['mri', 'm'],
+    ['mri', 'voxel'],
+    ['mri', 'mm'],
+    ['tkras', 'mm'],
+    ['mni', 'voxel'],
+    ['mni', 'mm'],
+    ['tkras', 'voxel'],
+])
+def test_convert_coordunits(_temp_bids_root, to_frame, to_unit):
     """Test conversion of coordinate units between voxel and xyz."""
+    print('INSIDE THE TEST')
+    print(os.getcwd())
+
     bids_path = _bids_path.copy().update(root=_temp_bids_root)
     subject = bids_path.subject
     coordsystem_fpath = bids_path.copy().update(suffix='coordsystem',
@@ -49,27 +62,33 @@ def test_convert_coordunits(_temp_bids_root, to_coord):
     img_fpath = sensors_mm.intended_for
 
     # check error on unexpected kwarg
-    if to_coord == 'm':
+    if to_unit == 'm':
+        sensors_m = convert_coord_units(sensors=sensors_mm,
+                                        to_unit=to_unit,
+                                        round=False)
+        np.testing.assert_array_almost_equal(
+            1000. * sensors_m.get_coords(),
+            sensors_mm.get_coords())
         with pytest.raises(ValueError,
                            match='Converting coordinates '
-                                 'to m is not accepted.'):
+                                 'to km is not accepted.'):
             convert_coord_units(sensors=sensors_mm,
-                                to_coord=to_coord)
+                                to_unit='km')
         return
-    elif to_coord in ['tkras', 'mni']:
+    elif to_frame in ['tkras', 'mni']:
         # conversion should not work if starting from xyz coords
         with pytest.raises(ValueError,
                            match='Converting coordinates '
                                  'to .* is not accepted.'):
             convert_coord_units(sensors=sensors_mm,
-                                to_coord=to_coord)
+                                to_unit=to_frame)
 
         # conversion should not work if trying to go to mm
         with pytest.raises(ValueError,
                            match='Converting coordinates '
                                  'to mm is not accepted.'):
             convert_coord_space(sensors=sensors_mm,
-                                to_coord='mm')
+                                to_frame='mm')
 
         # conversion should not work if trying to start
         # from mm xyz coords
@@ -77,15 +96,33 @@ def test_convert_coordunits(_temp_bids_root, to_coord):
                            match='Converting coordinates '
                                  'requires sensor'):
             convert_coord_space(sensors=sensors_mm,
-                                to_coord='tkras')
-    assert sensors_mm.coord_unit == 'mm'
+                                to_frame='tkras')
 
-    # convert to voxel should just require convert elec coords
-    if to_coord == 'mri':
+    # default settings for the read in coordinates file
+    assert sensors_mm.coord_unit == 'mm'
+    assert sensors_mm.coord_system == 'mri'
+
+    # test units
+    if to_unit == 'mm':
         # convert sensors
         sensors_conv = convert_coord_units(sensors=sensors_mm,
-                                           to_coord=to_coord,
+                                           to_unit=to_unit,
                                            round=False)
+        np.testing.assert_array_almost_equal(sensors_conv.get_coords(),
+                                             sensors_mm.get_coords())
+    elif to_unit == 'voxel':
+        pass
+
+    # convert to voxel should just require convert elec coords
+    if to_frame == 'mri':
+        # convert sensors first to voxels
+        sensors_conv = convert_coord_units(sensors=sensors_mm,
+                                           to_unit='voxel',
+                                           round=False)
+
+        # returning nothing should be the same
+        sensors_mm_mri = convert_coord_space(sensors_mm,
+                                             to_frame=to_frame)
 
         # apply affine yourself to go from mm -> voxels
         img = nb.load(img_fpath)
@@ -94,42 +131,39 @@ def test_convert_coordunits(_temp_bids_root, to_coord):
 
         # round trip should be the same
         sensors_mm_new = convert_coord_units(sensors=sensors_conv,
-                                             to_coord='mm',
+                                             to_unit='mm',
                                              round=False)
         # the coordinates should match
         np.testing.assert_array_almost_equal(sensors_mm_new.get_coords(),
                                              sensors_mm.get_coords())
-    elif to_coord == 'mm':
-        # convert sensors
-        sensors_conv = convert_coord_units(sensors=sensors_mm,
-                                           to_coord=to_coord,
-                                           round=False)
 
-        coords = sensors_mm.get_coords()
-    elif to_coord == 'mni':
+    elif to_frame == 'mni':
         # convert to voxels
         sensors_vox = convert_coord_units(sensors=sensors_mm,
-                                          to_coord='mri',
+                                          to_unit='voxel',
                                           round=False)
         # convert voxels to mni
-        sensors_conv = convert_coord_space(sensors_vox, to_coord=to_coord)
+        sensors_conv = convert_coord_space(sensors_vox, to_frame=to_frame,
+                                           subjects_dir=subjects_dir)
 
         # load FreeSurfer -> MNI transform (i.e. fsaverage)
-        mni_mri_t = read_talxfm(subject=subject, subjects_dir=subjects_dir)
+        mni_mri_t = read_talxfm(subject=f'sub-{subject}',
+                                subjects_dir=subjects_dir)
 
         # go voxels -> tkras
-        coords = apply_affine(mni_mri_t, sensors_vox.get_coords())
+        coords = apply_affine(mni_mri_t['trans'] * 1000.,
+                              sensors_vox.get_coords())
 
         # the coordinates should match
         np.testing.assert_array_almost_equal(sensors_conv.get_coords(),
                                              coords)
-    elif to_coord == 'tkras':
+    elif to_frame == 'tkras':
         # convert to voxels
         sensors_vox = convert_coord_units(sensors=sensors_mm,
-                                          to_coord='mri',
+                                          to_unit='voxel',
                                           round=False)
         # convert voxels to tkras
-        sensors_conv = convert_coord_space(sensors_vox, to_coord=to_coord)
+        sensors_conv = convert_coord_space(sensors_vox, to_frame=to_frame)
 
         # load FreeSurfer MGH file
         img = nb.load(T1mgz)
@@ -145,10 +179,16 @@ def test_convert_coordunits(_temp_bids_root, to_coord):
     assert img_fpath == sensors_conv.intended_for
 
     # new coordinate unit should be set
-    assert sensors_conv.coord_unit == to_coord
+    if to_frame == 'tkras':
+        sensors_conv.coord_unit == 'mm'
+    else:
+        assert sensors_conv.coord_unit == 'voxel'
+    assert sensors_conv.coord_system == to_frame
 
     # the coordinates should match
     np.testing.assert_array_equal(coords, sensors_conv.get_coords())
     assert all([sensors_conv.__dict__[key] == sensors_mm.__dict__[key]
                 for key in
-                sensors_conv.__dict__.keys() if key not in ['x', 'y', 'z', 'coord_unit']])
+                sensors_conv.__dict__.keys() if key not in ['x', 'y', 'z',
+                                                            'coord_unit',
+                                                            'coord_system']])